Numerous stability and safety improvements
  * Test for existence of ~/code before creating instead of
    trying with error redirected to /dev/null
  * Allow user to specify the `-j` option to make, defaulting to 4
  * Copy OS X application bundle to /Applications with build-date
    prefix instead of attempting to move the existing Emacs.app,
    which may require root permissions
  * Create symbolic links to native-lisp and lisp directories in
    the new Emacs-${BUILD}.app bundle
  * Replace symbolic links to emacs and emacsclient in /usr/local/bin to point
    to the newly built binaries

+++ build.sh
@@ -1,12 +1,20 @@
 #!/usr/bin/env bash
 
+set -euo pipefail
+
+# User Parameters
+# TODO: Default to all cores
+N_CORES=${1:-4}
 
 BUILD=$(date "+%Y-%m-%d")
 
 ROOT="$HOME/code"
 REPO="${ROOT}/emacs-native"
 
-mkdir "$ROOT" >/dev/null 2>/dev/null
+if [ ! -d "$ROOT" ]; then
+    echo "Creating ${ROOT}"
+    mkdir "$ROOT"
+fi
 
 if [ ! -d "$REPO" ]; then
     git clone --depth 1 git@github.com:emacs-mirror/emacs.git "$REPO" -b feature/native-comp
@@ -14,6 +22,7 @@ else
     cd $REPO && git pull
 fi
 
+# TODO: Make gcc location user-configurable
 export PATH="/usr/local/opt/gnu-sed/libexec/gnubin:${PATH}"
 export CFLAGS="-I/usr/local/Cellar/gcc/10.2.0/include -O3 -march=native"
 export LDFLAGS="-L/usr/local/Cellar/gcc/10.2.0/lib/gcc/10 -I/usr/local/Cellar/gcc/10.2.0/include"
@@ -44,28 +53,27 @@ git clean -xfd
      --with-tiff
 
 # comp-speed 3 may introduce dangerous optimizations
-make -j4 NATIVE_FULL_AOT=1 BYTE_COMPILE_EXTRA_FLAGS='--eval "(setq comp-speed 2)"'
+make -j${N_CORES} NATIVE_FULL_AOT=1 BYTE_COMPILE_EXTRA_FLAGS='--eval "(setq comp-speed 2)"'
 
 make install-eln
 make install
 
-cd /usr/local/bin || exit
-rm emacs
-rm emacsclient
-ln -s /usr/local/opt/gccemacs-${BUILD}/bin/emacs .
-ln -s /usr/local/opt/gccemacs-${BUILD}/bin/emacsclient .
-
-# Rename existing Emacs.app
-APP = "/Applications/Emacs.app"
-if [ -d "$APP" ]; then
-    mv $APP "/Applications/Emacs-$BUILD.app"
-fi
-
 # Move Emacs.app to /Applications
 if [ -d "${REPO}/nextstep/Emacs.app" ]; then
-    mv "${REPO}/nextstep/Emacs.app" "/Applications/"
+    cp -a "${REPO}/nextstep/Emacs.app" "/Applications/Emacs-${BUILD}.app"
 
-    cd /Applications/Emacs.app/Contents || exit
+    cd /Applications/Emacs-${BUILD}.app/Contents || exit
     ln -s /usr/local/opt/gccemacs-${BUILD}/lib/emacs/28.0.50/native-lisp .
     ln -s /usr/local/opt/gccemacs-${BUILD}/share/emacs/28.0.50/lisp .
+fi
+
+# Replace symbolic links in /usr/local/bin
+NEW_BIN_DIR="/usr/local/opt/gccemacs-${BUILD}/bin"
+if [ -e "$NEW_BIN_DIR" ]; then
+    echo "Replacing symbolic links in /usr/local/bin"
+    cd /usr/local/bin || exit
+    rm emacs
+    rm emacsclient
+    ln -s "${NEW_BIN_DIR}/emacs" .
+    ln -s "${NEW_BIN_DIR}/emacsclient" .
 fi
